Replacing CustomCardContainer.tsx with CustomCardHelper.tsx
